[libgens] Vdp::Reset(): Use an int for the for loop instead of size_t.
size_t is 64-bit on amd64, which requires a REX prefix. We only need
to iterate over 24 elements, so we don't need a 64-bit counter.

+++ src/libgens/Vdp/VdpIo.cpp
@@ -115,7 +115,7 @@ void Vdp::Reset(void)
 		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
 	};
 	
-	for (size_t i = 0; i < sizeof(vdp_reg_init)/sizeof(vdp_reg_init[0]); i++)
+	for (int i = 0; i < (int)(sizeof(vdp_reg_init)/sizeof(vdp_reg_init[0])); i++)
 	{
 		Set_Reg(i, vdp_reg_init[i]);
 	}
